requirements.txt
python-telegram-bot==13.7
requests
argparse
Filters
yt-dlp
ffmpeg